app/requirements.txt
Django==3.0.7
gunicorn==20.0.4
psycopg2-binary==2.8.5
factory-boy==2.12.0
faker==2.0.4
